+++ src/main/java/otakuplus/straybird/othellogame/network/http/HttpRequestUtil.java
@@ -17,15 +17,15 @@ import java.util.List;
 public class HttpRequestUtil {
     public static final String HOST_BASE_URL = "http://localhost:8080";
 
-    public static HttpRequest buildHttpGetRequest(String url){
+    public static HttpRequest buildHttpGetRequest(String url) {
         HttpRequestFactory requestFactory = HttpRequestFactorySingleton.getHttpRequestFactoryInstance();
         GenericUrl genericUrl = new GenericUrl(url);
         HttpRequest request = null;
         ApplicationContext applicationContext = ApplicationContextSingleton.getInstance();
         try {
             request = requestFactory.buildGetRequest(genericUrl);
-            if(applicationContext.getCurrentCookie() != null) {
-                System.out.println("current Cookie: "+applicationContext.getCurrentCookie());
+            if (applicationContext.getCurrentCookie() != null) {
+                System.out.println("current Cookie: " + applicationContext.getCurrentCookie());
                 request.getHeaders().set("cookie", applicationContext.getCurrentCookie());
                 List<HttpCookie> cookie = HttpCookie.parse(applicationContext.getCurrentCookie().get(0));
                 request.getHeaders().set("X-XSRF-TOKEN", cookie.get(0).getValue());
@@ -36,16 +36,16 @@ public class HttpRequestUtil {
         return request;
     }
 
-    public static HttpRequest buildHttpPostRequest(String url, Object object){
+    public static HttpRequest buildHttpPostRequest(String url, Object object) {
         HttpRequestFactory requestFactory = HttpRequestFactorySingleton.getHttpRequestFactoryInstance();
         GenericUrl genericUrl = new GenericUrl(url);
         HttpRequest request = null;
-        JsonHttpContent jsonHttpContent = new JsonHttpContent(JsonFactorySingleton.getJsonFactoryInstance() ,object);
+        JsonHttpContent jsonHttpContent = new JsonHttpContent(JsonFactorySingleton.getJsonFactoryInstance(), object);
         ApplicationContext applicationContext = ApplicationContextSingleton.getInstance();
         try {
             request = requestFactory.buildPostRequest(genericUrl, jsonHttpContent);
-            if(applicationContext.getCurrentCookie() != null) {
-                System.out.println("current Cookie: "+applicationContext.getCurrentCookie());
+            if (applicationContext.getCurrentCookie() != null) {
+                System.out.println("current Cookie: " + applicationContext.getCurrentCookie());
                 request.getHeaders().set("cookie", applicationContext.getCurrentCookie());
                 List<HttpCookie> cookie = HttpCookie.parse(applicationContext.getCurrentCookie().get(0));
                 request.getHeaders().set("X-XSRF-TOKEN", cookie.get(0).getValue());
@@ -56,15 +56,15 @@ public class HttpRequestUtil {
         return request;
     }
 
-    public static User getUserByHref(String href){
+    public static User getUserByHref(String href) {
         User user = null;
-        if(href != null){
+        if (href != null) {
             HttpRequest request = HttpRequestUtil.buildHttpGetRequest(href);
             HttpResponse response;
             try {
-                response =  request.execute();
+                response = request.execute();
                 user = response.parseAs(User.class);
-                if(user != null){
+                if (user != null) {
                     System.out.println("UserInformation: " + user.getLinks().getUserInformation().getHref());
                 }
             } catch (IOException e) {
@@ -74,43 +74,43 @@ public class HttpRequestUtil {
         return user;
     }
 
-    public static UserInformation getUserInformationByHref(String href){
+    public static UserInformation getUserInformationByHref(String href) {
         UserInformation userInformation = null;
-        if(href != null){
+        if (href != null) {
             HttpRequest request = HttpRequestUtil.buildHttpGetRequest(href);
             HttpResponse response;
-            try{
+            try {
                 response = request.execute();
                 userInformation = response.parseAs(UserInformation.class);
-                if(userInformation !=null){
+                if (userInformation != null) {
                     System.out.println("UserNickName " + userInformation.getNickname());
                 }
-            }catch (IOException e){
+            } catch (IOException e) {
                 e.printStackTrace();
             }
         }
         return userInformation;
     }
 
-    public static void updateUserOnlineList(){
+    public static void updateUserOnlineList() {
         String url = HttpRequestUtil.HOST_BASE_URL
-                +"/api/userOnlines/search/findByOnlineState?onlineState="
+                + "/api/userOnlines/search/findByOnlineState?onlineState="
                 + UserOnline.ONLINE;
         HttpResponse response = null;
         HttpRequest request;
         ApplicationContext applicationContext = ApplicationContextSingleton.getInstance();
-        try{
+        try {
             request = HttpRequestUtil.buildHttpGetRequest(url);
             response = request.execute();
 
-            if(response!= null && response.getStatusCode() == HttpStatusCodes.STATUS_CODE_OK){
+            if (response != null && response.getStatusCode() == HttpStatusCodes.STATUS_CODE_OK) {
                 EmBeddedUserOnlineList embeddedUserOnlineList = response.parseAs(EmBeddedUserOnlineList.class);
-                if(embeddedUserOnlineList != null){
+                if (embeddedUserOnlineList != null) {
                     UserOnlineList userOnlineList = embeddedUserOnlineList.getUserOnlineList();
-                    if(userOnlineList != null && userOnlineList.getUserOnlines() != null){
+                    if (userOnlineList != null && userOnlineList.getUserOnlines() != null) {
                         applicationContext.getUserInformationList().clear();
                         ArrayList<UserOnline> userOnlines = userOnlineList.getUserOnlines();
-                        for (int i=0; i<userOnlines.size();i++){
+                        for (int i = 0; i < userOnlines.size(); i++) {
                             User user = HttpRequestUtil.getUserByHref(userOnlines.get(i).getLinks().getUser().getHref());
                             UserInformation userInformation = HttpRequestUtil.getUserInformationByHref(user.getLinks().getUserInformation().getHref());
                             applicationContext.getUserInformationList().add(userInformation);
@@ -118,29 +118,29 @@ public class HttpRequestUtil {
                     }
                 }
             }
-        } catch (IOException e){
+        } catch (IOException e) {
             e.printStackTrace();
         }
     }
 
-    public static void updateGameTableList(){
-        String url = HttpRequestUtil.HOST_BASE_URL+"/api/gameTables";
+    public static void updateGameTableList() {
+        String url = HttpRequestUtil.HOST_BASE_URL + "/api/gameTables";
         HttpResponse response = null;
         HttpRequest request;
         ApplicationContext applicationContext = ApplicationContextSingleton.getInstance();
-        try{
+        try {
             request = HttpRequestUtil.buildHttpGetRequest(url);
             response = request.execute();
 
-            if(response!= null && response.getStatusCode() == HttpStatusCodes.STATUS_CODE_OK){
+            if (response != null && response.getStatusCode() == HttpStatusCodes.STATUS_CODE_OK) {
                 EmBeddedGameTableList embeddedGameTableList = response.parseAs(EmBeddedGameTableList.class);
-                if(embeddedGameTableList != null){
+                if (embeddedGameTableList != null) {
                     GameTableList gameTableList = embeddedGameTableList.getGameTableList();
-                    if(gameTableList != null && gameTableList.getGameTables() != null){
+                    if (gameTableList != null && gameTableList.getGameTables() != null) {
                         applicationContext.getGameTableList().clear();
                         ArrayList<GameTable> gameTables = gameTableList.getGameTables();
-                        for (int i=0; i<gameTables.size();i++){
-                            gameTables.get(i).setGameTableId(new Long(i+1));
+                        for (int i = 0; i < gameTables.size(); i++) {
+                            gameTables.get(i).setGameTableId(new Integer(i + 1));
                             User user = HttpRequestUtil.getUserByHref(gameTables.get(i)
                                     .getLinks().getPlayerA().getHref());
                             gameTables.get(i).setPlayerA(user);
@@ -152,7 +152,7 @@ public class HttpRequestUtil {
                     }
                 }
             }
-        } catch (IOException e){
+        } catch (IOException e) {
             e.printStackTrace();
         }
     }
